requirements.txt
llama-index
llama-index-readers-docling
llama-index-embeddings-huggingface
transformers
tqdm
requests
numpy